+++ pkg/main/repository.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 import (
 	"database/sql"
@@ -11,7 +11,6 @@ type MysqlRepository struct {
 }
 
 func NewMysqlRepository(config *Config) (mysqlRepository *MysqlRepository) {
-
 	return &MysqlRepository{connectDb(config.GetDbConnectionString())}
 }
 
@@ -26,6 +25,7 @@ type FeedRss struct {
 func (feed FeedRss) Url() string {
 	return feed.url
 }
+
 func (feed FeedRss) Id() int {
 	return feed.id
 }
@@ -88,6 +88,7 @@ func (repo *MysqlRepository) FindFeedRssIdByUrl(url string) (id int) {
 	}
 	return id
 }
+
 func (repo *MysqlRepository) AddFeedRssVisited(feed_id int, updated *time.Time) {
 	stmt, err := repo.db.Prepare("INSERT INTO feed_rss_visited(feed_id,updated) VALUES(?,?)")
 	if err != nil {
@@ -113,6 +114,7 @@ func (repo *MysqlRepository) GetLastFeedRssUpdatedByFeedId(id int) (t time.Time)
 
 	return t
 }
+
 func (repo MysqlRepository) AddShortlink(id string, url string) error {
 	stmt, err := repo.db.Prepare("INSERT INTO shortlink(uuid, url) VALUES(?,?)")
 	if err != nil {
